rename sdk-go/openedge to package openedge
