+++ master/api/server.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 
 	"github.com/baidu/openedge/protocol/http"
-	"github.com/baidu/openedge/sdk-go/openedge"
+	openedge "github.com/baidu/openedge/sdk/openedge-go"
 	"github.com/baidu/openedge/utils"
 )
