+++ master/engine/docker/instance.go
@@ -3,7 +3,7 @@ package docker
 import (
 	"github.com/baidu/openedge/logger"
 	"github.com/baidu/openedge/master/engine"
-	"github.com/baidu/openedge/sdk-go/openedge"
+	openedge "github.com/baidu/openedge/sdk/openedge-go"
 	"github.com/baidu/openedge/utils"
 )
