+++ master/engine/docker/service.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 
 	"github.com/baidu/openedge/logger"
-	"github.com/baidu/openedge/sdk-go/openedge"
+	openedge "github.com/baidu/openedge/sdk/openedge-go"
 	"github.com/orcaman/concurrent-map"
 )
