+++ master/engine/native/instance.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/baidu/openedge/logger"
 	"github.com/baidu/openedge/master/engine"
-	"github.com/baidu/openedge/sdk-go/openedge"
+	openedge "github.com/baidu/openedge/sdk/openedge-go"
 	"github.com/baidu/openedge/utils"
 )
