+++ master/engine/native/process.go
@@ -6,7 +6,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/baidu/openedge/sdk-go/openedge"
+	openedge "github.com/baidu/openedge/sdk/openedge-go"
 )
 
 type processConfigs struct {
