+++ master/engine/native/service.go
@@ -7,7 +7,7 @@ import (
 	"sync"
 
 	"github.com/baidu/openedge/logger"
-	"github.com/baidu/openedge/sdk-go/openedge"
+	openedge "github.com/baidu/openedge/sdk/openedge-go"
 	"github.com/orcaman/concurrent-map"
 )
