+++ master/engine/supervisor.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/baidu/openedge/logger"
-	"github.com/baidu/openedge/sdk-go/openedge"
+	openedge "github.com/baidu/openedge/sdk/openedge-go"
 	"github.com/jpillora/backoff"
 )
