+++ master/service.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 
 	"github.com/baidu/openedge/master/engine"
-	"github.com/baidu/openedge/sdk-go/openedge"
+	openedge "github.com/baidu/openedge/sdk/openedge-go"
 	"github.com/docker/distribution/uuid"
 )
