+++ openedge-agent/download.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"path"
 
-	"github.com/baidu/openedge/sdk-go/openedge"
+	openedge "github.com/baidu/openedge/sdk/openedge-go"
 	"github.com/baidu/openedge/utils"
 	"github.com/baidubce/bce-sdk-go/http"
 	"github.com/mholt/archiver"
