+++ openedge-function-manager/function.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/baidu/openedge/logger"
-	"github.com/baidu/openedge/sdk-go/openedge"
+	openedge "github.com/baidu/openedge/sdk/openedge-go"
 	"github.com/baidu/openedge/utils"
 	pool "github.com/jolestar/go-commons-pool"
 )
