+++ openedge-function-manager/instance.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"os"
 
-	"github.com/baidu/openedge/sdk-go/openedge"
+	openedge "github.com/baidu/openedge/sdk/openedge-go"
 )
 
 // Instance function instance interface
