+++ openedge-hub/main.go
@@ -8,7 +8,7 @@ import (
 	"github.com/baidu/openedge/openedge-hub/rule"
 	"github.com/baidu/openedge/openedge-hub/server"
 	"github.com/baidu/openedge/openedge-hub/session"
-	"github.com/baidu/openedge/sdk-go/openedge"
+	openedge "github.com/baidu/openedge/sdk/openedge-go"
 	"github.com/baidu/openedge/utils"
 )
