+++ sdk/openedge-go/function_client.go
