+++ sdk/openedge-go/generate.go
