+++ sdk/openedge-go/inspect.go
